Headers Reorganization and Public/Private funtions
So now the BMP_XX doesn't include any headers
this labour is delegate to BMP.c; I think that's
way is easier to read and expand.

Welcome to public and private funtions!
So the functions having a BMP_ as prefix are meant
to be use by users and those starting with and _
are meant to be called by the public ones so the
code is easier to read and expand.